Fixed up the messages app models
I removed some redundant inherits from the Chat and Message model, UUIDModel already inherits from models.Model.

Changed the `is_read` method on Chat to be a property, this can help with caching and also makes the api similar to Message's `is_read`.

Removed 'accounts.Users' and replaced it with a call to `get_user_model` so if the model ever changes our code won't break.